refactor(air): add cse module and make evaluate_on_random_inputs private
